Register a custom shard indexer for the dagstore to recognize data segment indices when present.
This change will result in lotus parsing deals for local indexing according to https://github.com/filecoin-project/FIPs/discussions/512

This is a continuation of https://github.com/filecoin-project/dagstore/pull/154 - This piece of the logic was seen as filecoin-specific rather than part of common dagstore functionality.
A parallel to this logic is expected to be ported also to boost, but there remain confguration in the use of the dagstore where this is the common locations where this indexing
functionality can be registered.